requirements.txt
absl-py==2.2.1
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==4.2.2
candlelite==1.0.17
Candlestick==0.0.8
catboost==1.2.7
ccxt==4.4.98
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
deap==1.4.2
distro==1.9.0
docutils==0.21.2
exceptiongroup==1.2.2
Farama-Notifications==0.0.4
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.2.2
grpcio==1.71.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.27.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaitaistruct==0.10
keras==3.9.2
keyboard==0.13.5
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kivymd==1.2.0
kiwisolver==1.4.7
libclang==18.1.1
lightgbm==4.6.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.3.2
namex==0.0.8
narwhals==1.34.1
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
okx==2.1.2
openai==1.97.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
optuna==4.3.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.2
paux==1.0.14
pefile==2023.2.7
pendulum==3.1.0
pillow==11.1.0
plotly==6.0.1
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycares==4.5.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pydeck==0.9.1
pydivert==2.1.0
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyMsgBox==1.0.9
pynput==1.8.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.15
pytweening==1.2.0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
requests==2.32.3
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.24.0
sb3_contrib==2.6.0
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.31.0
selenium-wire==5.1.0
Shimmy==2.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
stable_baselines3==2.6.0
starlette==0.46.1
stopit==1.1.2
streamlit==1.44.1
sympy==1.14.0
ta==0.11.0
TA-Lib @ file:///C:/Users/Administrator/Downloads/TA_Lib-0.4.28-cp39-cp39-win_amd64.whl#sha256=97020bb479143af698c2689b94de4a5b2651be685f27a39f5263084d2c2be41e
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.0
threadpoolctl==3.6.0
toml==0.10.2
torch==2.7.0
tornado==6.1
TPOT==0.12.2
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
update-checker==0.18.0
urllib3==1.26.12
uvicorn==0.34.0
watchdog==6.0.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
winloop==0.1.8
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.1.4
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
